[refactor] #35 상수 한곳에서 관리
