+++ core/designsystem/src/main/java/com/example/designsystem/component/textfield/DPlayTextArea.kt
@@ -29,8 +29,8 @@ import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.dplay.designsystem.R
-import com.example.designsystem.component.textfield.constant.TextFieldConstant
 import com.example.designsystem.theme.DPlayTheme
+import com.example.designsystem.util.TextFieldConstant
 
 private const val TEXT_AREA_ASPECT_RATIO = 343f / 180f
