--- core/designsystem/src/main/java/com/example/designsystem/component/textfield/constant/TextFieldConstant.kt
@@ -1,8 +1,0 @@
-package com.example.designsystem.component.textfield.constant
-
-object TextFieldConstant {
-    const val MIN_NICKNAME_LENGTH = 2
-    const val MAX_NICKNAME_LENGTH = 10
-
-    const val MAX_COMMENT_LENGTH = 150
-}
